requirements.txt
azure-identity
azure-storage-blob
azure-data-tables